backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.9
requests==2.32.3
python-dotenv==1.0.1
